#refactor 회원가입 코드 리펙토링
## 🏷 작업 분류

- [ ]  기능 추가
- [x]  리팩토링
- [ ]  버그 수정
- [ ]  환경 수정 및 기타

## 💡 작업 개요
회원가입 코드 위치 바꾸고 리펙토링했습니다.
## ***📜** 리뷰 요청사항*
